optimized the re-establishing of endpoint sockets down to <2 secs
identified the cause behind sleep() vs accept() bottleneck (previous commits).
and solved it by introducing binary backoff to old (dead) connections.
this may cause trouble down the line for critical re-attempts of connections.
this is a tradeoff to busy waiting of dead/useless endpoints that are
only kept around to see if they re-connect
also added locking to socket handling in class methods