bugfix: MAD-based threshold computes the correct values
MadTM was working utterly, wrong, as the function of MAD in
anomaly detection was falsely interpreted and thus implemented.
As a result, the threshold (which is fixed in this approach) was
computed based on a malformed vector of the z-scores of the
current sample window of the data stream.

This has now been fixed, and the documentation has been adjusted
accordingly. However, we might want to investigate, whether this
approach should even be represented in this module. It uses a
fixed threshold and not even an actual model and also requires
a static window to compute the median --- which is also not
suited to the use case!