Feature: Events and Labeling support (#31)
Added support for labeling arbitrary data points in dictionary format by
implementing events. The events are handled by the event handler, which
supports adding new events, parsing conditions for the event to create
an evaluation function, and labeling data points using the events.
A new library, pyparsing, is used for parsing the conditions for the
events. The code has been tested, but still requires in-depth tests in
the future, because of the parser.

---------

Signed-off-by: Fabian Sebastian Hofmann <fabian@alberts-hofmann.de>
Co-authored-by: Fabian Sebastian Hofmann <fabian@alberts-hofmann.de>
Co-authored-by: Nattacker <137180058+nattacker@users.noreply.github.com>